+++ index.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="icon" href="/images/favicon/favicon.ico">
-    <link rel="stylesheet" href="/style/reset.css" />
-    <link rel="stylesheet" href="/style/index.css" />
+    <link rel="stylesheet" href="/styles/reset.css" />
+    <link rel="stylesheet" href="/styles/index.css" />
     <title>Notion Project.</title>
 </head>
 <body>
